% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbiclean.R
\name{gbiclean}
\alias{gbiclean}
\title{GBI clean}
\usage{
gbiclean(x)
}
\arguments{
\item{input}{df, or tbl a dataframe}
}
\value{
smaller df or tbl
}
\description{
tidies and selects some variables from GBIF.org data
}
\examples{
gbiclean()
}
\seealso{
\code{\link[=gbifread]{gbifread()}}
}
\keyword{select}
\keyword{tidy,}
