% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbidata.r
\docType{data}
\name{gbidata}
\alias{gbidata}
\title{GBIF Data}
\format{
a data frame with 15 observations and 50 variables
Note: the first column has a quotation mark; this is an artifact of creating a smaller file in excel
It does not appear in the original data.
}
\source{
GBIF.org (27 October 2020) GBIF Occurrence Download /url{https://doi.org/10.15468/dl.24n5p9}
}
\usage{
gbidata
}
\description{
Part of a dataset from the Global Biodiversity Information Facility
of occurrence records of Vanessa cardui between 1990-1999
Full dataset has 33206 observations
}
\keyword{datasets}
