% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idbdata.R
\docType{data}
\name{idbdata}
\alias{idbdata}
\title{iDigBio data}
\format{
a data frame with 15 observations and 184 variables
}
\source{
/url{http://www.idigbio.org/portal} (2020),
3427 records, accessed on 2020-10-27T19:38:47.660127
}
\usage{
idbdata
}
\description{
Part of a dataset from iDigBio.org
of occurrence records of Vanessa cardui
Full dataset has 3427 observations
}
\keyword{datasets}
