% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idigclean.R
\name{idigclean}
\alias{idigclean}
\title{Clean iDigBio Data}
\usage{
idigclean(x)
}
\arguments{
\item{input}{tbl_df a dataframe}
}
\value{
smaller tbl_df
}
\description{
Tidies and selects some variables from iDigBio data
}
\section{Warning:}{
this function must be used after \code{\link[=rmcolon]{rmcolon()}}
}

\examples{
 
idigclean()

}
\seealso{
\code{\link[=idigfread]{idigfread()}} \code{\link[=rmcolon]{rmcolon()}}
}
\keyword{select}
\keyword{tidy,}
