% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idigfread.r
\name{idigfread}
\alias{idigfread}
\title{Read iDigBio}
\usage{
idigfread(x)
}
\arguments{
\item{input}{a comma-delimited csv}

\item{output}{tb df a tibble or dataframe}
}
\description{
reads a csv file from iDigBio.org
}
\examples{
idigfread() 
}
\keyword{csv}
\keyword{parse,}
