% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmcolon.R
\name{rmcolon}
\alias{rmcolon}
\title{Remove Colon}
\usage{
rmcolon(df)
}
\arguments{
\item{input}{df, a dataframe}

\item{output}{df, the dataframe}
}
\description{
replaces colon with dot in column names
}
\examples{
rmcolon()
}
\keyword{header,}
\keyword{replacement}
