% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbifread.R
\name{gbifread}
\alias{gbifread}
\title{Read GBIF Raw Data}
\usage{
gbifread(x)
}
\arguments{
\item{input}{a tab-delimited csv}
}
\value{
tbl_df a tibble or dataframe
}
\description{
reads a csv file from GBIF.org with \code{\link[=fread]{fread()}}
}
\examples{
gbifread("tsvfilename.csv")
}
\keyword{parse,}
\keyword{tsv}
