% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idigfread.r
\name{idigfread}
\alias{idigfread}
\title{Read iDigBio Raw Data}
\usage{
idigfread(x)
}
\arguments{
\item{input}{a comma-delimited csv}

\item{output}{tb df a tibble or dataframe}
}
\description{
reads a csv file from iDigBio.org with \code{\link[=fread]{fread()}}
}
\examples{
idigfread("filename.csv") 
}
\keyword{csv}
\keyword{parse,}
